+++ tests/unit/test_handler_creator.py
@@ -8,9 +8,9 @@ from bigxml.handler_marker import xml_handle_element, xml_handle_text
 from bigxml.nodes import XMLElement, XMLElementAttributes, XMLText
 
 
-def create_nodes(*path):
+def create_nodes(*path, parent=None):
     # create nodes
-    nodes = []
+    nodes = [parent] if parent is not None else []
     for node_name in path:
         if node_name == ":text:":
             node = XMLText(text="text", parents=tuple(nodes))
@@ -23,14 +23,28 @@ def create_nodes(*path):
         nodes.append(node)
 
     # set handles
-    for i, parent in enumerate(nodes):
-        if not isinstance(parent, XMLElement):
+    for i, node_parent in enumerate(nodes):
+        if not isinstance(node_parent, XMLElement):
             continue
+
+        # get existing children
         try:
-            child = nodes[i + 1]
-            parent.set_handle(lambda h, child=child: h(child))
+            children = list(node_parent.iter_from(lambda n: (n,)))
+        except RuntimeError:
+            children = []
+
+        # append new children
+        try:
+            children.append(nodes[i + 1])
         except IndexError:
-            parent.set_handle(lambda h: ())
+            pass
+
+        # create handle
+        def handle(handler, _children):
+            for child in _children:
+                yield from handler(child)
+
+        node_parent.set_handle(lambda h, c=children: handle(h, c))
 
     return nodes
 
@@ -191,7 +205,7 @@ def test_syntactic_sugar_two_levels(test_create_handler, handler):
 
 
 #
-# class instance
+# class instance & class (generic case)
 #
 
 
@@ -211,7 +225,8 @@ def test_syntactic_sugar_two_levels(test_create_handler, handler):
     (("t0", "t1", ":text:"), "5", ":text:"),
     (("t0", "t1", ":text:", "t2"), "5", ":text:"),
 )
-def test_class_instance(test_create_handler):
+@pytest.mark.parametrize("instantiate_class", (False, True))
+def test_class_instance(test_create_handler, instantiate_class):
     class Handler:
         @xml_handle_element("a")
         @staticmethod
@@ -244,7 +259,8 @@ def test_class_instance(test_create_handler):
         def handle5(node):
             yield ("5", node)
 
-    test_create_handler(Handler())
+    handler = Handler() if instantiate_class else Handler
+    test_create_handler(handler)
 
 
 @cases(
@@ -255,7 +271,8 @@ def test_class_instance(test_create_handler):
     (("x1", "y1", "z", "a"), "0", "a"),
     (("x1", "y1", "z", "w", "a"), None, None),
 )
-def test_marked_class_instance(test_create_handler):
+@pytest.mark.parametrize("instantiate_class", (False, True))
+def test_marked_class_instance(test_create_handler, instantiate_class):
     @xml_handle_element("x0", "y0")
     @xml_handle_element("x1", "y1", "z")
     class Handler:
@@ -264,7 +281,8 @@ def test_marked_class_instance(test_create_handler):
         def handle0(node):
             yield ("0", node)
 
-    test_create_handler(Handler())
+    handler = Handler() if instantiate_class else Handler
+    test_create_handler(handler)
 
 
 @cases(
@@ -280,7 +298,8 @@ def test_marked_class_instance(test_create_handler):
     (("x", "y1", "z"), None, None),
     (("x", "y1", "z", "a"), None, None),
 )
-def test_deep_marked_class_instances(test_create_handler):
+@pytest.mark.parametrize("instantiate_class", (False, True))
+def test_deep_marked_class_instances(test_create_handler, instantiate_class):
     @xml_handle_element("y0")
     class DeepHandler0:
         @xml_handle_element("a")
@@ -302,7 +321,117 @@ def test_deep_marked_class_instances(test_create_handler):
         def __init__(self):
             self.deep_handler1 = DeepHandler1()
 
-    test_create_handler(Handler())
+    handler = Handler() if instantiate_class else Handler
+    test_create_handler(handler)
+
+
+#
+# class
+#
+
+
+def test_class_init():
+    @xml_handle_element("x", "y")
+    class Handler:
+        init_nb = 0
+
+        def __init__(self, node):
+            self.init_nb = Handler.init_nb
+            Handler.init_nb += 1
+            self.root = node
+            self.seen_nodes = 0
+
+        @xml_handle_element("a", "b")
+        def handle0(self, node):
+            yield (self.init_nb, self.root, self.seen_nodes, node)
+            self.seen_nodes += 1
+
+    #   x -> y0 -> a0 -> b0
+    #                 -> b1
+    #           -> a1 -> b2
+    #     -> y1 -> a2 -> b3
+    #
+    # Handler should be instanciated on y0 and y1
+    #
+    # the use of namespaces below is to avoid e.g. node_y0==node_y1
+    #
+    # pylint: disable=unbalanced-tuple-unpacking
+    node_x, node_y0 = create_nodes("x", "{y0}y")
+    _, node_a0, node_b0 = create_nodes("{a0}a", "{b0}b", parent=node_y0)
+    _, node_b1 = create_nodes("{b1}b", parent=node_a0)
+    _, _, node_b2 = create_nodes("{a1}a", "{b2}b", parent=node_y0)
+    _, node_y1, _, node_b3 = create_nodes("{n1}y", "{a2}a", "{b3}b", parent=node_x)
+    # pylint: enable=unbalanced-tuple-unpacking
+
+    handler = create_handler(Handler)
+    assert list(handler(node_x)) == [
+        (0, node_y0, 0, node_b0),
+        (0, node_y0, 1, node_b1),
+        (0, node_y0, 2, node_b2),
+        (1, node_y1, 0, node_b3),
+    ]
+    assert Handler.init_nb == 2
+
+
+def test_class_init_no_mandatory_parameters():
+    @xml_handle_element("x")
+    class Handler:
+        def __init__(self, answer=42):
+            self.answer = answer
+
+        @xml_handle_element("a")
+        def handle0(self, node):
+            yield ("0", self.answer, node)
+
+    nodes = create_nodes("x", "a")
+    handler = create_handler(Handler)
+    assert list(handler(nodes[0])) == [("0", 42, nodes[1])]
+
+
+def test_class_init_one_mandatory_parameter():
+    @xml_handle_element("x")
+    class Handler:
+        def __init__(self, node, answer=42):
+            self.node = node
+            self.answer = answer
+
+        @xml_handle_element("a")
+        def handle0(self, node):
+            yield ("0", self.node, self.answer, node)
+
+    nodes = create_nodes("x", "a")
+    handler = create_handler(Handler)
+    assert list(handler(nodes[0])) == [("0", nodes[0], 42, nodes[1])]
+
+
+def test_class_init_two_mandatory_parameters():
+    @xml_handle_element("x")
+    class Handler:
+        def __init__(self, node, answer):
+            self.node = node
+            self.answer = answer
+
+    nodes = create_nodes("x", "a")
+    handler = create_handler(Handler)
+    with pytest.raises(TypeError) as excinfo:
+        list(handler(nodes[0]))
+
+    assert "__init__ should have" in str(excinfo.value)
+    assert "node, answer" in str(excinfo.value)
+
+
+def test_class_init_crash():
+    @xml_handle_element("x")
+    class Handler:
+        def __init__(self):
+            raise TypeError("Something went wrong")
+
+    nodes = create_nodes("x", "a")
+    handler = create_handler(Handler)
+    with pytest.raises(TypeError) as excinfo:
+        list(handler(nodes[0]))
+
+    assert str(excinfo.value) == "Something went wrong"
 
 
 #
@@ -343,7 +472,6 @@ def test_several_catchall_handlers():
         b"a",
         {"a", "b"},
         {"a": lambda _: None},
-        object,
         object(),
     ),
     ids=type,